fixed scanner numbers

+++ asg2/lyutils.h
@@ -20,7 +20,7 @@ extern FILE* yyin;
 extern char* yytext; 
 extern int yy_flex_debug;
 extern int yydebug;
-extern int yyleng; 
+extern size_t yyleng; 
 
 int yylex();
 int yylex_destroy();
